+++ clang-tools-extra/clang-tidy/modernize/UseIntegerSignComparisonCheck.cpp
@@ -80,11 +80,13 @@ UseIntegerSignComparisonCheck::UseIntegerSignComparisonCheck(
     : ClangTidyCheck(Name, Context),
       IncludeInserter(Options.getLocalOrGlobal("IncludeStyle",
                                                utils::IncludeSorter::IS_LLVM),
-                      areDiagsSelfContained()) {}
+                      areDiagsSelfContained()),
+      ConsideringIntSize(Options.get("ConsideringIntegerSize", true)) {}
 
 void UseIntegerSignComparisonCheck::storeOptions(
     ClangTidyOptions::OptionMap &Opts) {
   Options.store(Opts, "IncludeStyle", IncludeInserter.getStyle());
+  Options.store(Opts, "ConsideringIntegerSize", ConsideringIntSize);
 }
 
 void UseIntegerSignComparisonCheck::registerMatchers(MatchFinder *Finder) {
@@ -139,10 +141,11 @@ void UseIntegerSignComparisonCheck::check(
   if (LHS == nullptr || RHS == nullptr)
     return;
 
-  if (Result.Context->getTypeSize(
-          SignedCastExpression->getSubExpr()->getType()) >
-      Result.Context->getTypeSize(
-          UnSignedCastExpression->getSubExpr()->getType()))
+  if (ConsideringIntSize &&
+      (Result.Context->getTypeSize(
+           SignedCastExpression->getSubExpr()->getType()) >
+       Result.Context->getTypeSize(
+           UnSignedCastExpression->getSubExpr()->getType())))
     return;
 
   const Expr *SubExprLHS = nullptr;